Added Copy Records Functionality
With Copy Records Functionality now it's possible to fully copy a record that has already been saved in the database. This allows the application to maintain records history and to copy the entire diet from one Customer to another.

A class called CopyFactory was created to automatically copy properties states from one object to another of the same class. Now it's no longer necessary to call the .findLast() method of DAO's classes to update the state of an object. When the method .save() is used by an object, if no error happens, the object updates it's own states, including the new id if it is a new record.